third_party/WebKit/LayoutTests/fast/dom/DeviceMotion/create-event-expected.txt

Tests that document.createEvent() works with DeviceMotionEvent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof event == 'object' is true
PASS 'type' in event is true
PASS 'bubbles' in event is true
PASS 'cancelable' in event is true
FAIL 'acceleration' in event should be true. Was false.
FAIL 'accelerationIncludingGravity' in event should be true. Was false.
FAIL 'rotationRate' in event should be true. Was false.
FAIL 'interval' in event should be true. Was false.
PASS typeof event.type == 'string' is true
PASS typeof event.bubbles == 'boolean' is true
PASS typeof event.cancelable == 'boolean' is true
FAIL typeof event.acceleration == 'object' should be true. Was false.
FAIL typeof event.accelerationIncludingGravity == 'object' should be true. Was false.
FAIL typeof event.rotationRate == 'object' should be true. Was false.
FAIL typeof event.interval == 'object' should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
